Fixed an issue with not all stats from equips resetting on equip,  Fixed an issue where secret item image wouldn't laod properly. Changed main.py to the name of the game
